Abstracted to grid environment, nArmBandit completely broken now :P

+++ src/main/kotlin/Environment/Point.kt
@@ -1,6 +1,5 @@
 package Environment
 
 
-class Point (vararg val dimensions : Double){
-
+class Point (vararg val dimensions : Int){
 }
